Przeniesienie logiki z dto do serwisu
Dodanie fontu jako resource
Stworzenie stałych
